test: clear the remaining /W4 warnings; assert previously-dropped load results
Finishes the lint pass: tests/ now compiles warning-free at /W4, so the whole
codebase is clean and a NEW warning stands out instead of hiding in noise.

The most substantive change -- discarded [[nodiscard]] returns (C4834). These
split into two genuinely different cases and are treated differently:

  * Calls made FOR THEIR SIDE EFFECT -- reading the WD2793 status register acks
    INTRQ, reading the data register advances the sector byte stream. The value
    is deliberately unwanted, so an explicit (void) states that. (6 sites; one
    already carried the comment "ack the Restore's own INTRQ".)

  * SETUP calls that report success -- load_cartridge()/slot.load() return
    CartridgeLoadResult, NOT bool, and the tests were dropping it entirely. A
    failed load would have surfaced later as a puzzling assertion about memory
    contents rather than "the cartridge never loaded". These are now asserted
    as `== CartridgeLoadResult::Ok`, matching the convention the same files
    already use elsewhere. (15 sites.)

  memory_battery_backed_sram_unit_test is the one exception: it has no expect()
  (it uses `if (!expect_true(...)) return 1;`) and the call sits in a lambda
  returning a Sram, so it cannot signal failure that way. A failed save is
  already caught by the content comparison that follows, so it gets an honest
  (void) rather than a fake assertion.

Also:
  - C4100 (6): lambdas that must keep a parameter to match the callback
    signature now leave it UNNAMED, the idiomatic "required but unused".
    Line-targeted -- the same text appears in ~159 lambdas that DO use it.
  - C4456 (1): a second `id` shadowed the first; renamed to id2.
  - C4244: sector lists were initializer_list<unsigned> bound to uint8_t, and
    `0x10 + channel` promoted to int at a uint8_t parameter -- both spelled
    explicitly now.
  - C4244 (the last ones, reported deep inside MSVC's <utility>): build_probe
    took std::vector<std::pair<uint16_t,uint8_t>>, whose constructor narrows
    int literals. Replaced with a small ProbeWrite aggregate -- braced init
    from FITTING integer constants is not a narrowing conversion, so every
    `{{0x4000, 3}}` call site is unchanged, and addr/value read better than
    first/second.

Gates: build clean at /W4 with ZERO warnings from src/ or tests/; ctest
-LE slow_cpu_sweep 273/273, both game-asset oracles running for real.

+++ tests/unit/frontend/machine_audio_mixer_fm_unit_test.cpp
@@ -109,9 +109,9 @@ void program_fm_tone_channel(Ym2413Opll& opll, const int channel) {
     write_reg(opll, 0x05, 0xF0);  // car AR=15 (instant)
     write_reg(opll, 0x06, 0x00);
     write_reg(opll, 0x07, 0x00);
-    write_reg(opll, 0x10 + channel, 0x00);
-    write_reg(opll, 0x30 + channel, 0x00);  // user patch, volume 0
-    write_reg(opll, 0x20 + channel, 0x19);  // key-on, block 4, fnum 256
+    write_reg(opll, static_cast<std::uint8_t>(0x10 + channel), 0x00);
+    write_reg(opll, static_cast<std::uint8_t>(0x30 + channel), 0x00);  // user patch, volume 0
+    write_reg(opll, static_cast<std::uint8_t>(0x20 + channel), 0x19);  // key-on, block 4, fnum 256
 }
 
 // The literal zero-FM presenter/mixer arithmetic, reproduced independently:
